more refactoring for clean arch
