+++ src/main/java/data_access/DataAccess.java
@@ -1,4 +1,4 @@
-package use_case.search;
+package data_access;
 import java.io.BufferedReader;
 import java.io.InputStreamReader;
 import java.net.HttpURLConnection;
