+++ src/main/java/interface_adapter/search/GameSearchPresenter.java
@@ -1,4 +1,4 @@
-package interface_adapter;
+package interface_adapter.search;
 
 import view.GameSearchView;
