--- src/main/java/interface_adapter/search/SearchController.java
@@ -1,5 +1,0 @@
-package interface_adapter.search;
-
-public class SearchController {
-    
-}
