--- src/main/java/use_case/search/SearchInputBoundary.java
@@ -1,6 +1,0 @@
-package use_case.search;
-
-public interface SearchInputBoundary {
-
-    void execute(SearchInputData searchInputData);
-}
