--- src/main/java/use_case/search/SearchOutputData.java
@@ -1,5 +1,0 @@
-package use_case.search;
-
-public class SearchOutputData {
-
-}
