--- src/main/java/use_case/search/SearchUserDataAccessInterface.java
@@ -1,4 +1,0 @@
-package use_case.search;
-
-public interface SearchUserDataAccessInterface {
-}
